removed event callbacks, fixed typing bugs

+++ src/orrery/components/business.py
@@ -408,7 +408,7 @@ class Unemployed(StatusComponent):
 
     __slots__ = "years"
 
-    def __init__(self, created: str, years: int = 0.0) -> None:
+    def __init__(self, created: str, years: float = 0.0) -> None:
         super().__init__(created)
         self.years: float = years
